Resolve remaining test failures
