+++ tests/integration/test_groupby.py
@@ -2,6 +2,22 @@ import dask.dataframe as dd
 import numpy as np
 import pandas as pd
 import pytest
+from dask.datasets import timeseries
+
+
+@pytest.fixture()
+def timeseries_df(c):
+    pdf = timeseries(freq="1d").compute().reset_index(drop=True)
+
+    # input nans in pandas dataframe
+    col1_index = np.random.randint(0, 30, size=int(pdf.shape[0] * 0.2))
+    col2_index = np.random.randint(0, 30, size=int(pdf.shape[0] * 0.3))
+    pdf.loc[col1_index, "x"] = np.nan
+    pdf.loc[col2_index, "y"] = np.nan
+
+    c.create_table("timeseries", pdf, persist=True)
+
+    return None
 
 
 def test_group_by(c):
@@ -226,7 +242,7 @@ def test_aggregations(c):
     dd.assert_eq(return_df.reset_index(drop=True), expected_df)
 
 
-def test_stats_aggregation(c):
+def test_stats_aggregation(c, timeseries_df):
     # test regr_count
     regr_count = c.sql(
         """
@@ -293,7 +309,7 @@ def test_stats_aggregation(c):
         SELECT
             name,
             AVG(y) FILTER (WHERE x IS NOT NULL) as avg_y,
-            AVG(x) FILTER (WHERE x IS NOT NULL) as avg_x,
+            AVG(x) FILTER (WHERE x IS NOT NULL) as avg_x
         FROM timeseries
         GROUP BY name
     ) SELECT
@@ -320,7 +336,7 @@ def test_stats_aggregation(c):
         SELECT
             name,
             AVG(y) FILTER (WHERE x IS NOT NULL) as avg_y,
-            AVG(x) FILTER (WHERE x IS NOT NULL) as avg_x,
+            AVG(x) FILTER (WHERE x IS NOT NULL) as avg_x
         FROM timeseries
         GROUP BY name
     ) SELECT
